(#5) Отделение ВМ в качестве библиотеки
