--- .idea/runConfigurations/Release.xml
@@ -1,18 +1,0 @@
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="Release" type="CargoCommandRunConfiguration" factoryName="Cargo Command">
-    <option name="command" value="run --release" />
-    <option name="workingDirectory" value="file://$PROJECT_DIR$" />
-    <option name="channel" value="DEFAULT" />
-    <option name="requiredFeatures" value="true" />
-    <option name="allFeatures" value="false" />
-    <option name="emulateTerminal" value="false" />
-    <option name="withSudo" value="false" />
-    <option name="backtrace" value="SHORT" />
-    <envs />
-    <option name="isRedirectInput" value="false" />
-    <option name="redirectInputPath" value="" />
-    <method v="2">
-      <option name="CARGO.BUILD_TASK_PROVIDER" enabled="true" />
-    </method>
-  </configuration>
-</component>
