+++ .idea/runConfigurations/Runner.xml
@@ -1,6 +1,6 @@
 <component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="Debug" type="CargoCommandRunConfiguration" factoryName="Cargo Command">
-    <option name="command" value="run" />
+  <configuration default="false" name="Runner" type="CargoCommandRunConfiguration" factoryName="Cargo Command">
+    <option name="command" value="run --bin movm-runner examples/fibonacci.mobc" />
     <option name="workingDirectory" value="file://$PROJECT_DIR$" />
     <option name="channel" value="DEFAULT" />
     <option name="requiredFeatures" value="true" />
