+++ Cargo.toml
@@ -3,6 +3,10 @@ name = "movm"
 version = "0.1.0"
 edition = "2021"
 
+[lib]
+name = "movm"
+path = "src/movm/main.rs"
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
