+++ src/movm/main.rs
@@ -1,0 +1,4 @@
+pub mod inst;
+pub mod vm;
+pub mod loader;
+mod stack;
